fix tests + code cleanup
