+++ core/src/main/java/com/filmfit/core/Core.java
@@ -2,13 +2,18 @@ package com.filmfit.core;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
+import org.springframework.context.annotation.Import;
 import org.springframework.context.annotation.PropertySource;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.scheduling.annotation.EnableAsync;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
 
 @SpringBootApplication
 @EnableTransactionManagement
 @EnableAsync
+@EnableJpaRepositories
+@Import(DataSourceAutoConfiguration.class)
 @PropertySource("classpath:core.properties")
 public class Core {
