+++ core/src/main/java/com/filmfit/core/common/Company.java
@@ -1,13 +1,12 @@
 package com.filmfit.core.common;
 
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import jakarta.validation.constraints.NotBlank;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Entity
 @Data
@@ -24,5 +23,6 @@ public class Company {
 
     private String logoPath;
 
+    // TODO: change to Country
     private String originCountry;
 }
