+++ core/src/main/java/com/filmfit/core/common/Language.java
@@ -1,13 +1,12 @@
 package com.filmfit.core.common;
 
-import jakarta.persistence.Transient;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
+import jakarta.persistence.Transient;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Entity
 @Data
@@ -16,13 +15,11 @@ import jakarta.persistence.Id;
 @Builder
 public class Language {
 
-    @Id
-    private String iso6391;
-
-    private String name;
-
-    private String englishName;
-
     @Transient
     public static final Language NONE = new Language("xx", "No Language", "No Language");
+
+    @Id
+    private String iso6391;
+    private String name;
+    private String englishName;
 }
