+++ core/src/main/java/com/filmfit/core/film/Genre.java
@@ -1,19 +1,15 @@
 package com.filmfit.core.film;
 
 import jakarta.persistence.Column;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import lombok.AccessLevel;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import jakarta.validation.constraints.NotBlank;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 import lombok.NonNull;
-import lombok.RequiredArgsConstructor;
 import lombok.Setter;
 
 @Entity
@@ -28,7 +24,7 @@ public class Genre {
     @NonNull
     private Long id;
 
-    @Column(unique = true)
+    @Column(unique = true, updatable = false)
     @NotBlank
     @NonNull
     private String name;
