--- core/src/main/java/com/filmfit/core/film/ReactiveFilmRepository.java
@@ -1,6 +1,0 @@
-package com.filmfit.core.film;
-
-import org.springframework.data.repository.reactive.ReactiveCrudRepository;
-
-interface ReactiveFilmRepository extends ReactiveCrudRepository<Film, Long> {
-}
