+++ external/src/main/java/com/filmfit/external/tmdb/dto/FilmListItem.java
@@ -2,12 +2,15 @@ package com.filmfit.external.tmdb.dto;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import jakarta.validation.constraints.*;
-import org.hibernate.validator.constraints.Range;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Positive;
 import java.util.List;
+import org.hibernate.validator.constraints.Range;
 
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public record FilmListItem (
+public record FilmListItem(
     @NotNull @Positive Long id,
     @NotNull Boolean adult,
     String backdropPath,
@@ -22,5 +25,6 @@ public record FilmListItem (
     @NotNull Boolean video,
     @Range(min = 0, max = 10) Double voteAverage,
     @Positive Long voteCount
-) {}
+) {
+}
