+++ external/src/main/java/com/filmfit/external/tmdb/mapping/CountryMapper.java
@@ -4,7 +4,6 @@ import com.filmfit.core.common.Country;
 import com.filmfit.external.tmdb.dto.CountryInfo;
 import lombok.NonNull;
 import org.mapstruct.Mapper;
-import org.mapstruct.Mapping;
 
 @Mapper(componentModel = "spring")
 public interface CountryMapper {
