+++ external/src/main/resources/external.properties
@@ -1,14 +1,10 @@
 spring.application.name=external
-
 spring.jackson.deserialization.fail-on-unknown-properties=false
-
 spring.threads.virtual.enabled=true
-
 tmdb.base-url=https://api.themoviedb.org/3
 tmdb.endpoints.movies-discover=/discover/movie
 tmdb.endpoints.movies-details=/movie/{movieId}
 tmdb.endpoints.languages=/configuration/languages
 tmdb.endpoints.countries=/configuration/countries
 tmdb.endpoints.collections=/search/collection
-
 logging.level.com.filmfit.external.tmdb=TRACE
