+++ external/src/test/java/com/filmfit/external/ExternalApplicationTests.java
@@ -2,7 +2,6 @@ package com.filmfit.external;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.annotation.Import;
 
 @SpringBootTest
 class ExternalApplicationTests {
