+++ src/main/resources/application.properties
@@ -1,2 +1,1 @@
 spring.application.name=matching
-# logging.level.com.filmfit.external.tmdb=DEBUG
