+++ src/test/java/com/filmfit/matching/TestMatchingApplication.java
@@ -1,7 +1,9 @@
 package com.filmfit.matching;
 
 import org.springframework.boot.SpringApplication;
+import org.springframework.test.context.ContextConfiguration;
 
+@ContextConfiguration(classes = {TestMatchingApplication.class})
 public class TestMatchingApplication {
 
     public static void main(String[] args) {
